move all systems to their own files and removed magic numbers where possible. May move the magic numbers to a system assumtions json table later.
