+++ lib/openstudio-standards/standards/necb/NECB2011/hvac_systems.rb
@@ -1,4 +1,3 @@
-require_relative("autozone.rb")
 class NECB2011
   def model_add_hvac(model:)
     OpenStudio.logFree(OpenStudio::Info, 'openstudio.model.Model', 'Started Adding HVAC')
@@ -1359,1101 +1358,11 @@ class NECB2011
     return "NECB-CNEB ClimatZone 4-8"
   end
 
-  def add_sys1_unitary_ac_baseboard_heating(model:,
-                                            zones:,
-                                            mau_type:,
-                                            mau_heating_coil_type:,
-                                            baseboard_type:,
-                                            hw_loop:)
-    # System Type 1: PTAC with no heating (unitary AC)
-    # Zone baseboards, electric or hot water depending on argument baseboard_type
-    # baseboard_type choices are "Hot Water" or "Electric"
-    # PSZ to represent make-up air unit (if present)
-    # This measure creates:
-    # a PTAC  unit for each zone in the building; DX cooling coil
-    # and heating coil that is always off
-    # Baseboards ("Hot Water or "Electric") in zones connected to hot water loop
-    # MAU is present if argument mau == true, not present if argument mau == false
-    # MAU is PSZ; DX cooling
-    # MAU heating coil: hot water coil or electric, depending on argument mau_heating_coil_type
-    # mau_heating_coil_type choices are "Hot Water", "Electric"
-    # boiler_fueltype choices match OS choices for Boiler component fuel type, i.e.
-    # "NaturalGas","Electricity","PropaneGas","FuelOil#1","FuelOil#2","Coal","Diesel","Gasoline","OtherFuel1"
 
-    # Some system parameters are set after system is set up; by applying method 'apply_hvac_efficiency_standard'
 
-    always_on = model.alwaysOnDiscreteSchedule
-    always_off = BTAP::Resources::Schedules::StandardSchedules::ON_OFF.always_off(model)
 
-    # Create MAU
-    # TO DO: MAU sizing, characteristics (fan operation schedules, temperature setpoints, outdoor air, etc)
 
-    if mau_type == true
 
-      mau_air_loop = OpenStudio::Model::AirLoopHVAC.new(model)
-      mau_air_loop.setName('Sys_1_Make-up air unit')
-
-      # When an air_loop is constructed, its constructor creates a sizing:system object
-      # the default sizing:system constructor makes a system:sizing object
-      # appropriate for a multizone VAV system
-      # this systems is a constant volume system with no VAV terminals,
-      # and therfore needs different default settings
-      air_loop_sizing = mau_air_loop.sizingSystem # TODO units
-      air_loop_sizing.setTypeofLoadtoSizeOn('VentilationRequirement')
-      air_loop_sizing.autosizeDesignOutdoorAirFlowRate
-      air_loop_sizing.setMinimumSystemAirFlowRatio(1.0)
-      air_loop_sizing.setPreheatDesignTemperature(7.0)
-      air_loop_sizing.setPreheatDesignHumidityRatio(0.008)
-      air_loop_sizing.setPrecoolDesignTemperature(13.0)
-      air_loop_sizing.setPrecoolDesignHumidityRatio(0.008)
-      air_loop_sizing.setCentralCoolingDesignSupplyAirTemperature(13)
-      air_loop_sizing.setCentralHeatingDesignSupplyAirTemperature(43)
-      air_loop_sizing.setSizingOption('NonCoincident')
-      air_loop_sizing.setAllOutdoorAirinCooling(true)
-      air_loop_sizing.setAllOutdoorAirinHeating(true)
-      air_loop_sizing.setCentralCoolingDesignSupplyAirHumidityRatio(0.0085)
-      air_loop_sizing.setCentralHeatingDesignSupplyAirHumidityRatio(0.0080)
-      air_loop_sizing.setCoolingDesignAirFlowMethod('DesignDay')
-      air_loop_sizing.setCoolingDesignAirFlowRate(0.0)
-      air_loop_sizing.setHeatingDesignAirFlowMethod('DesignDay')
-      air_loop_sizing.setHeatingDesignAirFlowRate(0.0)
-      air_loop_sizing.setSystemOutdoorAirMethod('ZoneSum')
-
-      mau_fan = OpenStudio::Model::FanConstantVolume.new(model, always_on)
-
-      if mau_heating_coil_type == 'Electric' # electric coil
-        mau_htg_coil = OpenStudio::Model::CoilHeatingElectric.new(model, always_on)
-      end
-
-      if mau_heating_coil_type == 'Hot Water'
-        mau_htg_coil = OpenStudio::Model::CoilHeatingWater.new(model, always_on)
-        hw_loop.addDemandBranchForComponent(mau_htg_coil)
-      end
-
-      # Set up DX coil with default curves (set to NECB);
-
-      mau_clg_coil = self.add_onespeed_DX_coil(model, always_on)
-
-      # oa_controller
-      oa_controller = OpenStudio::Model::ControllerOutdoorAir.new(model)
-      oa_controller.autosizeMinimumOutdoorAirFlowRate
-      # oa_controller.setEconomizerControlType("DifferentialEnthalpy")
-
-      # oa_system
-      oa_system = OpenStudio::Model::AirLoopHVACOutdoorAirSystem.new(model, oa_controller)
-
-      # Add the components to the air loop
-      # in order from closest to zone to furthest from zone
-      supply_inlet_node = mau_air_loop.supplyInletNode
-      mau_fan.addToNode(supply_inlet_node)
-      mau_htg_coil.addToNode(supply_inlet_node)
-      mau_clg_coil.addToNode(supply_inlet_node)
-      oa_system.addToNode(supply_inlet_node)
-
-      # Add a setpoint manager to control the supply air temperature
-      sat = 20.0
-      sat_sch = OpenStudio::Model::ScheduleRuleset.new(model)
-      sat_sch.setName('Makeup-Air Unit Supply Air Temp')
-      sat_sch.defaultDaySchedule.setName('Makeup Air Unit Supply Air Temp Default')
-      sat_sch.defaultDaySchedule.addValue(OpenStudio::Time.new(0, 24, 0, 0), sat)
-      setpoint_mgr = OpenStudio::Model::SetpointManagerScheduled.new(model, sat_sch)
-      setpoint_mgr.addToNode(mau_air_loop.supplyOutletNode)
-
-    end # Create MAU
-
-
-    zones.each do |zone|
-      # Zone sizing temperature
-      sizing_zone = zone.sizingZone
-      sizing_zone.setZoneCoolingDesignSupplyAirTemperature(13.0)
-      sizing_zone.setZoneHeatingDesignSupplyAirTemperature(43.0)
-      sizing_zone.setZoneCoolingSizingFactor(1.1)
-      sizing_zone.setZoneHeatingSizingFactor(1.3)
-
-      # Create a PTAC for each zone:
-      # PTAC DX Cooling with electric heating coil; electric heating coil is always off
-      # TO DO: need to apply this system to space types:
-      # (1) data processing area: control room, data centre
-      # when cooling capacity <= 20kW and
-      # (2) residential/accommodation: murb, hotel/motel guest room
-      # when building/space heated only (this as per NECB; apply to
-      # all for initial work? CAN-QUEST limitation)
-
-      # TO DO: PTAC characteristics: sizing, fan schedules, temperature setpoints, interaction with MAU
-
-      # htg_coil_elec = OpenStudio::Model::CoilHeatingElectric.new(model,always_on)
-      add_ptac_dx_cooling(model, zone)
-
-      # add zone baseboards
-      add_zone_baseboards(baseboard_type: baseboard_type, hw_loop: hw_loop, model: model, zone: zone)
-
-      #  # Create a diffuser and attach the zone/diffuser pair to the MAU air loop, if applicable
-      if mau_type == true
-        diffuser = OpenStudio::Model::AirTerminalSingleDuctUncontrolled.new(model, always_on)
-        mau_air_loop.addBranchForZone(zone, diffuser.to_StraightComponent)
-      end # components for MAU
-    end # of zone loop
-
-    return true
-  end
-
-  # sys1_unitary_ac_baseboard_heating
-
-  def add_sys1_unitary_ac_baseboard_heating_multi_speed(model:,
-                                                        zones:,
-                                                        mau:,
-                                                        mau_heating_coil_type:,
-                                                        baseboard_type:,
-                                                        hw_loop:)
-    # System Type 1: PTAC with no heating (unitary AC)
-    # Zone baseboards, electric or hot water depending on argument baseboard_type
-    # baseboard_type choices are "Hot Water" or "Electric"
-    # PSZ to represent make-up air unit (if present)
-    # This measure creates:
-    # a PTAC  unit for each zone in the building; DX cooling coil
-    # and heating coil that is always off
-    # Baseboards ("Hot Water or "Electric") in zones connected to hot water loop
-    # MAU is present if argument mau == true, not present if argument mau == false
-    # MAU is PSZ; DX cooling
-    # MAU heating coil: hot water coil or electric, depending on argument mau_heating_coil_type
-    # mau_heating_coil_type choices are "Hot Water", "Electric"
-    # boiler_fueltype choices match OS choices for Boiler component fuel type, i.e.
-    # "NaturalGas","Electricity","PropaneGas","FuelOil#1","FuelOil#2","Coal","Diesel","Gasoline","OtherFuel1"
-
-    # Some system parameters are set after system is set up; by applying method 'apply_hvac_efficiency_standard'
-
-    always_on = model.alwaysOnDiscreteSchedule
-
-    # define always off schedule for ptac heating coil
-    always_off = BTAP::Resources::Schedules::StandardSchedules::ON_OFF.always_off(model)
-
-    # TODO: Heating and cooling temperature set point schedules are set somewhere else
-    # TODO: For now fetch the schedules and use them in setting up the heat pump system
-    # TODO: Later on these schedules need to be passed on to this method
-    htg_temp_sch = nil
-    clg_temp_sch = nil
-    zones.each do |izone|
-      if izone.thermostat.is_initialized
-        zone_thermostat = izone.thermostat.get
-        if zone_thermostat.to_ThermostatSetpointDualSetpoint.is_initialized
-          dual_thermostat = zone_thermostat.to_ThermostatSetpointDualSetpoint.get
-          htg_temp_sch = dual_thermostat.heatingSetpointTemperatureSchedule.get
-          clg_temp_sch = dual_thermostat.coolingSetpointTemperatureSchedule.get
-          break
-        end
-      end
-    end
-
-    # Create MAU
-    # TO DO: MAU sizing, characteristics (fan operation schedules, temperature setpoints, outdoor air, etc)
-
-    if mau == true
-
-      staged_thermostat = OpenStudio::Model::ZoneControlThermostatStagedDualSetpoint.new(model)
-      staged_thermostat.setHeatingTemperatureSetpointSchedule(htg_temp_sch)
-      staged_thermostat.setNumberofHeatingStages(4)
-      staged_thermostat.setCoolingTemperatureSetpointBaseSchedule(clg_temp_sch)
-      staged_thermostat.setNumberofCoolingStages(4)
-
-      mau_air_loop = OpenStudio::Model::AirLoopHVAC.new(model)
-
-      mau_air_loop.setName('Sys_1_Make-up air unit')
-
-      # When an air_loop is constructed, its constructor creates a sizing:system object
-      # the default sizing:system constructor makes a system:sizing object
-      # appropriate for a multizone VAV system
-      # this systems is a constant volume system with no VAV terminals,
-      # and therfore needs different default settings
-      air_loop_sizing = mau_air_loop.sizingSystem # TODO units
-      air_loop_sizing.setTypeofLoadtoSizeOn('Sensible')
-      air_loop_sizing.autosizeDesignOutdoorAirFlowRate
-      air_loop_sizing.setMinimumSystemAirFlowRatio(1.0)
-      air_loop_sizing.setPreheatDesignTemperature(7.0)
-      air_loop_sizing.setPreheatDesignHumidityRatio(0.008)
-      air_loop_sizing.setPrecoolDesignTemperature(12.8)
-      air_loop_sizing.setPrecoolDesignHumidityRatio(0.008)
-      air_loop_sizing.setCentralCoolingDesignSupplyAirTemperature(13.0)
-      air_loop_sizing.setCentralHeatingDesignSupplyAirTemperature(43.0)
-      air_loop_sizing.setSizingOption('NonCoincident')
-      air_loop_sizing.setAllOutdoorAirinCooling(false)
-      air_loop_sizing.setAllOutdoorAirinHeating(false)
-      air_loop_sizing.setCentralCoolingDesignSupplyAirHumidityRatio(0.0085)
-      air_loop_sizing.setCentralHeatingDesignSupplyAirHumidityRatio(0.0080)
-      air_loop_sizing.setCoolingDesignAirFlowMethod('DesignDay')
-      air_loop_sizing.setCoolingDesignAirFlowRate(0.0)
-      air_loop_sizing.setHeatingDesignAirFlowMethod('DesignDay')
-      air_loop_sizing.setHeatingDesignAirFlowRate(0.0)
-      air_loop_sizing.setSystemOutdoorAirMethod('ZoneSum')
-
-      mau_fan = OpenStudio::Model::FanConstantVolume.new(model, always_on)
-
-      # Multi-stage gas heating coil
-      if mau_heating_coil_type == 'Electric' || mau_heating_coil_type == 'Hot Water'
-
-        mau_htg_coil = OpenStudio::Model::CoilHeatingGasMultiStage.new(model)
-        mau_htg_stage_1 = OpenStudio::Model::CoilHeatingGasMultiStageStageData.new(model)
-        mau_htg_stage_2 = OpenStudio::Model::CoilHeatingGasMultiStageStageData.new(model)
-        mau_htg_stage_3 = OpenStudio::Model::CoilHeatingGasMultiStageStageData.new(model)
-        mau_htg_stage_4 = OpenStudio::Model::CoilHeatingGasMultiStageStageData.new(model)
-
-        if mau_heating_coil_type == 'Electric'
-
-          mau_supplemental_htg_coil = OpenStudio::Model::CoilHeatingElectric.new(model, always_on)
-
-        elsif mau_heating_coil_type == 'Hot Water'
-
-          mau_supplemental_htg_coil = OpenStudio::Model::CoilHeatingWater.new(model, always_on)
-          hw_loop.addDemandBranchForComponent(mau_supplemental_htg_coil)
-
-        end
-
-        mau_htg_stage_1.setNominalCapacity(0.1)
-        mau_htg_stage_2.setNominalCapacity(0.2)
-        mau_htg_stage_3.setNominalCapacity(0.3)
-        mau_htg_stage_4.setNominalCapacity(0.4)
-
-      end
-
-      # Add stages to heating coil
-      mau_htg_coil.addStage(mau_htg_stage_1)
-      mau_htg_coil.addStage(mau_htg_stage_2)
-      mau_htg_coil.addStage(mau_htg_stage_3)
-      mau_htg_coil.addStage(mau_htg_stage_4)
-
-      # TODO: other fuel-fired heating coil types? (not available in OpenStudio/E+ - may need to play with efficiency to mimic other fuel types)
-
-      # Set up DX cooling coil
-      mau_clg_coil = OpenStudio::Model::CoilCoolingDXMultiSpeed.new(model)
-      mau_clg_coil.setFuelType('Electricity')
-      mau_clg_stage_1 = OpenStudio::Model::CoilCoolingDXMultiSpeedStageData.new(model)
-      mau_clg_stage_2 = OpenStudio::Model::CoilCoolingDXMultiSpeedStageData.new(model)
-      mau_clg_stage_3 = OpenStudio::Model::CoilCoolingDXMultiSpeedStageData.new(model)
-      mau_clg_stage_4 = OpenStudio::Model::CoilCoolingDXMultiSpeedStageData.new(model)
-      mau_clg_coil.addStage(mau_clg_stage_1)
-      mau_clg_coil.addStage(mau_clg_stage_2)
-      mau_clg_coil.addStage(mau_clg_stage_3)
-      mau_clg_coil.addStage(mau_clg_stage_4)
-
-      air_to_air_heatpump = OpenStudio::Model::AirLoopHVACUnitaryHeatPumpAirToAirMultiSpeed.new(model, mau_fan, mau_htg_coil, mau_clg_coil, mau_supplemental_htg_coil)
-      #              air_to_air_heatpump.setName("#{zone.name} ASHP")
-      air_to_air_heatpump.setControllingZoneorThermostatLocation(zones[1])
-      air_to_air_heatpump.setSupplyAirFanOperatingModeSchedule(always_on)
-      air_to_air_heatpump.setNumberofSpeedsforHeating(4)
-      air_to_air_heatpump.setNumberofSpeedsforCooling(4)
-
-      # oa_controller
-      oa_controller = OpenStudio::Model::ControllerOutdoorAir.new(model)
-      oa_controller.autosizeMinimumOutdoorAirFlowRate
-      # oa_controller.setEconomizerControlType("DifferentialEnthalpy")
-
-      # oa_system
-      oa_system = OpenStudio::Model::AirLoopHVACOutdoorAirSystem.new(model, oa_controller)
-
-      # Add the components to the air loop
-      # in order from closest to zone to furthest from zone
-      supply_inlet_node = mau_air_loop.supplyInletNode
-      air_to_air_heatpump.addToNode(supply_inlet_node)
-      oa_system.addToNode(supply_inlet_node)
-
-    end # Create MAU
-
-
-
-    zones.each do |zone|
-      # Zone sizing temperature
-      sizing_zone = zone.sizingZone
-      sizing_zone.setZoneCoolingDesignSupplyAirTemperature(13.0)
-      sizing_zone.setZoneHeatingDesignSupplyAirTemperature(43.0)
-      sizing_zone.setZoneCoolingSizingFactor(1.1)
-      sizing_zone.setZoneHeatingSizingFactor(1.3)
-
-      # Set up PTAC heating coil; apply always off schedule
-
-      # htg_coil_elec = OpenStudio::Model::CoilHeatingElectric.new(model,always_on)
-      add_ptac_dx_cooling(model, zone)
-
-      # add zone baseboards
-      add_zone_baseboards(baseboard_type: baseboard_type, hw_loop: hw_loop, model: model, zone: zone)
-
-      #  # Create a diffuser and attach the zone/diffuser pair to the MAU air loop, if applicable
-      if mau == true
-
-        diffuser = OpenStudio::Model::AirTerminalSingleDuctUncontrolled.new(model, always_on)
-        mau_air_loop.addBranchForZone(zone, diffuser.to_StraightComponent)
-
-      end # components for MAU
-    end # of zone loop
-
-    return true
-  end
-
-  # sys1_unitary_ac_baseboard_heating
-
-  def add_sys2_FPFC_sys5_TPFC(model:,
-                              zones:,
-                              chiller_type:,
-                              fan_coil_type:,
-                              mau_cooling_type:,
-                              hw_loop:)
-    # System Type 2: FPFC or System 5: TPFC
-    # This measure creates:
-    # -a four pipe or a two pipe fan coil unit for each zone in the building;
-    # -a make up air-unit to provide ventilation to each zone;
-    # -a heating loop, cooling loop and condenser loop to serve four pipe fan coil units
-    # Arguments:
-    #   boiler_fueltype: "NaturalGas","Electricity","PropaneGas","FuelOil#1","FuelOil#2","Coal","Diesel","Gasoline","OtherFuel1"
-    #   chiller_type: "Scroll";"Centrifugal";"Rotary Screw";"Reciprocating"
-    #   mua_cooling_type: make-up air unit cooling type "DX";"Hydronic"
-    #   fan_coil_type options are "TPFC" or "FPFC"
-
-    # TODO: Add arguments as needed when the sizing routine is finalized. For example we will need to know the
-    # required size of the boilers to decide on how many units are needed based on NECB rules.
-
-    always_on = model.alwaysOnDiscreteSchedule
-
-    # schedule for two-pipe fan coil operation
-
-    twenty_four_hrs = OpenStudio::Time.new(0, 24, 0, 0)
-
-    # Heating coil availability schedule for tpfc
-    tpfc_htg_availability_sch = OpenStudio::Model::ScheduleRuleset.new(model)
-    tpfc_htg_availability_sch.setName('tpfc_htg_availability')
-    # Cooling coil availability schedule for tpfc
-    tpfc_clg_availability_sch = OpenStudio::Model::ScheduleRuleset.new(model)
-    tpfc_clg_availability_sch.setName('tpfc_clg_availability')
-    istart_month = [1, 7, 11]
-    istart_day = [1, 1, 1]
-    iend_month = [6, 10, 12]
-    iend_day = [30, 31, 31]
-    sch_htg_value = [1, 0, 1]
-    sch_clg_value = [0, 1, 0]
-    for i in 0..2
-      tpfc_htg_availability_sch_rule = OpenStudio::Model::ScheduleRule.new(tpfc_htg_availability_sch)
-      tpfc_htg_availability_sch_rule.setName('tpfc_htg_availability_sch_rule')
-      tpfc_htg_availability_sch_rule.setStartDate(model.getYearDescription.makeDate(istart_month[i], istart_day[i]))
-      tpfc_htg_availability_sch_rule.setEndDate(model.getYearDescription.makeDate(iend_month[i], iend_day[i]))
-      tpfc_htg_availability_sch_rule.setApplySunday(true)
-      tpfc_htg_availability_sch_rule.setApplyMonday(true)
-      tpfc_htg_availability_sch_rule.setApplyTuesday(true)
-      tpfc_htg_availability_sch_rule.setApplyWednesday(true)
-      tpfc_htg_availability_sch_rule.setApplyThursday(true)
-      tpfc_htg_availability_sch_rule.setApplyFriday(true)
-      tpfc_htg_availability_sch_rule.setApplySaturday(true)
-      day_schedule = tpfc_htg_availability_sch_rule.daySchedule
-      day_schedule.setName('tpfc_htg_availability_sch_rule_day')
-      day_schedule.addValue(twenty_four_hrs, sch_htg_value[i])
-
-      tpfc_clg_availability_sch_rule = OpenStudio::Model::ScheduleRule.new(tpfc_clg_availability_sch)
-      tpfc_clg_availability_sch_rule.setName('tpfc_clg_availability_sch_rule')
-      tpfc_clg_availability_sch_rule.setStartDate(model.getYearDescription.makeDate(istart_month[i], istart_day[i]))
-      tpfc_clg_availability_sch_rule.setEndDate(model.getYearDescription.makeDate(iend_month[i], iend_day[i]))
-      tpfc_clg_availability_sch_rule.setApplySunday(true)
-      tpfc_clg_availability_sch_rule.setApplyMonday(true)
-      tpfc_clg_availability_sch_rule.setApplyTuesday(true)
-      tpfc_clg_availability_sch_rule.setApplyWednesday(true)
-      tpfc_clg_availability_sch_rule.setApplyThursday(true)
-      tpfc_clg_availability_sch_rule.setApplyFriday(true)
-      tpfc_clg_availability_sch_rule.setApplySaturday(true)
-      day_schedule = tpfc_clg_availability_sch_rule.daySchedule
-      day_schedule.setName('tpfc_clg_availability_sch_rule_day')
-      day_schedule.addValue(twenty_four_hrs, sch_clg_value[i])
-
-    end
-
-    # Create a chilled water loop
-
-    chw_loop = OpenStudio::Model::PlantLoop.new(model)
-    chiller1, chiller2 = self.setup_chw_loop_with_components(model, chw_loop, chiller_type)
-
-    # Create a condenser Loop
-
-    cw_loop = OpenStudio::Model::PlantLoop.new(model)
-    ctower = self.setup_cw_loop_with_components(model, cw_loop, chiller1, chiller2)
-
-    # Set up make-up air unit for ventilation
-    # TO DO: Need to investigate characteristics of make-up air unit for NECB reference
-    # and define them here
-
-    air_loop = OpenStudio::Model::AirLoopHVAC.new(model)
-
-    air_loop.setName('Sys_2_Make-up air unit')
-
-    # When an air_loop is contructed, its constructor creates a sizing:system object
-    # the default sizing:system constructor makes a system:sizing object
-    # appropriate for a multizone VAV system
-    # this systems is a constant volume system with no VAV terminals,
-    # and therfore needs different default settings
-    air_loop_sizing = air_loop.sizingSystem # TODO units
-    air_loop_sizing.setTypeofLoadtoSizeOn('Sensible')
-    air_loop_sizing.autosizeDesignOutdoorAirFlowRate
-    air_loop_sizing.setMinimumSystemAirFlowRatio(1.0)
-    air_loop_sizing.setPreheatDesignTemperature(7.0)
-    air_loop_sizing.setPreheatDesignHumidityRatio(0.008)
-    air_loop_sizing.setPrecoolDesignTemperature(13.0)
-    air_loop_sizing.setPrecoolDesignHumidityRatio(0.008)
-    air_loop_sizing.setCentralCoolingDesignSupplyAirTemperature(13.0)
-    air_loop_sizing.setCentralHeatingDesignSupplyAirTemperature(13.1)
-    air_loop_sizing.setSizingOption('NonCoincident')
-    air_loop_sizing.setAllOutdoorAirinCooling(false)
-    air_loop_sizing.setAllOutdoorAirinHeating(false)
-    air_loop_sizing.setCentralCoolingDesignSupplyAirHumidityRatio(0.008)
-    air_loop_sizing.setCentralHeatingDesignSupplyAirHumidityRatio(0.008)
-    air_loop_sizing.setCoolingDesignAirFlowMethod('DesignDay')
-    air_loop_sizing.setCoolingDesignAirFlowRate(0.0)
-    air_loop_sizing.setHeatingDesignAirFlowMethod('DesignDay')
-    air_loop_sizing.setHeatingDesignAirFlowRate(0.0)
-    air_loop_sizing.setSystemOutdoorAirMethod('ZoneSum')
-
-    fan = OpenStudio::Model::FanConstantVolume.new(model, always_on)
-
-    # Assume direct-fired gas heating coil for now; need to add logic
-    # to set up hydronic or electric coil depending on proposed?
-
-    htg_coil = OpenStudio::Model::CoilHeatingGas.new(model, always_on)
-
-    # Add DX or hydronic cooling coil
-    if mau_cooling_type == 'DX'
-      clg_coil = self.add_onespeed_DX_coil(model, tpfc_clg_availability_sch)
-    elsif mau_cooling_type == 'Hydronic'
-      clg_coil = OpenStudio::Model::CoilCoolingWater.new(model, tpfc_clg_availability_sch)
-      chw_loop.addDemandBranchForComponent(clg_coil)
-    end
-
-    # does MAU have an economizer?
-    oa_controller = OpenStudio::Model::ControllerOutdoorAir.new(model)
-    oa_controller.autosizeMinimumOutdoorAirFlowRate
-
-    # oa_system = OpenStudio::Model::AirLoopHVACOutdoorAirSystem.new(model,oa_controller)
-    oa_system = OpenStudio::Model::AirLoopHVACOutdoorAirSystem.new(model, oa_controller)
-
-    # Add the components to the air loop
-    # in order from closest to zone to furthest from zone
-    supply_inlet_node = air_loop.supplyInletNode
-    fan.addToNode(supply_inlet_node)
-    htg_coil.addToNode(supply_inlet_node)
-    clg_coil.addToNode(supply_inlet_node)
-    oa_system.addToNode(supply_inlet_node)
-
-    # Add a setpoint manager single zone reheat to control the
-    # supply air temperature based on the needs of default zone (OpenStudio picks one)
-    # TO DO: need to have method to pick appropriate control zone?
-
-    setpoint_mgr_single_zone_reheat = OpenStudio::Model::SetpointManagerSingleZoneReheat.new(model)
-    setpoint_mgr_single_zone_reheat.setMinimumSupplyAirTemperature(13.0)
-    setpoint_mgr_single_zone_reheat.setMaximumSupplyAirTemperature(13.1)
-    setpoint_mgr_single_zone_reheat.addToNode(air_loop.supplyOutletNode)
-
-    # Set up FC (ZoneHVAC,cooling coil, heating coil, fan) in each zone
-
-    zones.each do |zone|
-      # Zone sizing temperature
-      sizing_zone = zone.sizingZone
-      sizing_zone.setZoneCoolingDesignSupplyAirTemperature(13.0)
-      sizing_zone.setZoneHeatingDesignSupplyAirTemperature(43.0)
-      sizing_zone.setZoneCoolingSizingFactor(1.1)
-      sizing_zone.setZoneHeatingSizingFactor(1.3)
-
-      # fc supply fan
-      fc_fan = OpenStudio::Model::FanConstantVolume.new(model, always_on)
-
-      if fan_coil_type == 'FPFC'
-        # heating coil
-        fc_htg_coil = OpenStudio::Model::CoilHeatingWater.new(model, always_on)
-
-        # cooling coil
-        fc_clg_coil = OpenStudio::Model::CoilCoolingWater.new(model, always_on)
-      elsif fan_coil_type == 'TPFC'
-        # heating coil
-        fc_htg_coil = OpenStudio::Model::CoilHeatingWater.new(model, tpfc_htg_availability_sch)
-
-        # cooling coil
-        fc_clg_coil = OpenStudio::Model::CoilCoolingWater.new(model, tpfc_clg_availability_sch)
-      end
-
-      # connect heating coil to hot water loop
-      hw_loop.addDemandBranchForComponent(fc_htg_coil)
-      # connect cooling coil to chilled water loop
-      chw_loop.addDemandBranchForComponent(fc_clg_coil)
-
-      zone_fc = OpenStudio::Model::ZoneHVACFourPipeFanCoil.new(model, always_on, fc_fan, fc_clg_coil, fc_htg_coil)
-      zone_fc.addToThermalZone(zone)
-
-      # Create a diffuser and attach the zone/diffuser pair to the air loop (make-up air unit)
-      diffuser = OpenStudio::Model::AirTerminalSingleDuctUncontrolled.new(model, always_on)
-      air_loop.addBranchForZone(zone, diffuser.to_StraightComponent)
-    end # zone loop
-  end
-
-  # add_sys2_FPFC_sys5_TPFC
-
-  def add_sys3and8_single_zone_packaged_rooftop_unit_with_baseboard_heating_single_speed(model:,
-                                                                                         zones:,
-                                                                                         heating_coil_type:,
-                                                                                         baseboard_type:,
-                                                                                         hw_loop:)
-    # System Type 3: PSZ-AC
-    # This measure creates:
-    # -a constant volume packaged single-zone A/C unit
-    # for each zone in the building; DX cooling with
-    # heating coil: fuel-fired or electric, depending on argument heating_coil_type
-    # heating_coil_type choices are "Electric", "Gas", "DX"
-    # zone baseboards: hot water or electric, depending on argument baseboard_type
-    # baseboard_type choices are "Hot Water" or "Electric"
-    # boiler_fueltype choices match OS choices for Boiler component fuel type, i.e.
-    # "NaturalGas","Electricity","PropaneGas","FuelOil#1","FuelOil#2","Coal","Diesel","Gasoline","OtherFuel1"
-
-    always_on = model.alwaysOnDiscreteSchedule
-
-    zones.each do |zone|
-      air_loop = OpenStudio::Model::AirLoopHVAC.new(model)
-
-      air_loop.setName("Sys_3_PSZ #{zone.name}")
-
-      # When an air_loop is constructed, its constructor creates a sizing:system object
-      # the default sizing:system constructor makes a system:sizing object
-      # appropriate for a multizone VAV system
-      # this systems is a constant volume system with no VAV terminals,
-      # and therfore needs different default settings
-      air_loop_sizing = air_loop.sizingSystem # TODO units
-      air_loop_sizing.setTypeofLoadtoSizeOn('Sensible')
-      air_loop_sizing.autosizeDesignOutdoorAirFlowRate
-      air_loop_sizing.setMinimumSystemAirFlowRatio(1.0)
-      air_loop_sizing.setPreheatDesignTemperature(7.0)
-      air_loop_sizing.setPreheatDesignHumidityRatio(0.008)
-      air_loop_sizing.setPrecoolDesignTemperature(13.0)
-      air_loop_sizing.setPrecoolDesignHumidityRatio(0.008)
-      air_loop_sizing.setCentralCoolingDesignSupplyAirTemperature(13.0)
-      air_loop_sizing.setCentralHeatingDesignSupplyAirTemperature(43)
-      air_loop_sizing.setSizingOption('NonCoincident')
-      air_loop_sizing.setAllOutdoorAirinCooling(false)
-      air_loop_sizing.setAllOutdoorAirinHeating(false)
-      air_loop_sizing.setCentralCoolingDesignSupplyAirHumidityRatio(0.0085)
-      air_loop_sizing.setCentralHeatingDesignSupplyAirHumidityRatio(0.0080)
-      air_loop_sizing.setCoolingDesignAirFlowMethod('DesignDay')
-      air_loop_sizing.setCoolingDesignAirFlowRate(0.0)
-      air_loop_sizing.setHeatingDesignAirFlowMethod('DesignDay')
-      air_loop_sizing.setHeatingDesignAirFlowRate(0.0)
-      air_loop_sizing.setSystemOutdoorAirMethod('ZoneSum')
-
-      # Zone sizing temperature
-      sizing_zone = zone.sizingZone
-      sizing_zone.setZoneCoolingDesignSupplyAirTemperature(13.0)
-      sizing_zone.setZoneHeatingDesignSupplyAirTemperature(43.0)
-      sizing_zone.setZoneCoolingSizingFactor(1.1)
-      sizing_zone.setZoneHeatingSizingFactor(1.3)
-
-      fan = OpenStudio::Model::FanConstantVolume.new(model, always_on)
-
-      case heating_coil_type
-      when 'Electric' # electric coil
-        htg_coil = OpenStudio::Model::CoilHeatingElectric.new(model, always_on)
-
-      when 'Gas'
-        htg_coil = OpenStudio::Model::CoilHeatingGas.new(model, always_on)
-
-      when 'DX'
-        htg_coil = OpenStudio::Model::CoilHeatingDXSingleSpeed.new(model)
-        supplemental_htg_coil = OpenStudio::Model::CoilHeatingElectric.new(model, always_on)
-        htg_coil.setMinimumOutdoorDryBulbTemperatureforCompressorOperation(-10.0)
-        sizing_zone.setZoneHeatingSizingFactor(1.3)
-        sizing_zone.setZoneCoolingSizingFactor(1.0)
-      else
-        raise("#{heating_coil_type} is not a valid heating coil type.)")
-      end
-
-      # TO DO: other fuel-fired heating coil types? (not available in OpenStudio/E+ - may need to play with efficiency to mimic other fuel types)
-
-      # Set up DX coil with NECB performance curve characteristics;
-      clg_coil = OpenStudio::Model::CoilCoolingDXSingleSpeed.new(model)
-
-      # oa_controller
-      oa_controller = OpenStudio::Model::ControllerOutdoorAir.new(model)
-      oa_controller.autosizeMinimumOutdoorAirFlowRate
-
-      # oa_system
-      oa_system = OpenStudio::Model::AirLoopHVACOutdoorAirSystem.new(model, oa_controller)
-
-      # Add the components to the air loop
-      # in order from closest to zone to furthest from zone
-      supply_inlet_node = air_loop.supplyInletNode
-      #              fan.addToNode(supply_inlet_node)
-      #              supplemental_htg_coil.addToNode(supply_inlet_node) if heating_coil_type == "DX"
-      #              htg_coil.addToNode(supply_inlet_node)
-      #              clg_coil.addToNode(supply_inlet_node)
-      #              oa_system.addToNode(supply_inlet_node)
-      if heating_coil_type == 'DX'
-        air_to_air_heatpump = OpenStudio::Model::AirLoopHVACUnitaryHeatPumpAirToAir.new(model, always_on, fan, htg_coil, clg_coil, supplemental_htg_coil)
-        air_to_air_heatpump.setName("#{zone.name} ASHP")
-        air_to_air_heatpump.setControllingZone(zone)
-        air_to_air_heatpump.setSupplyAirFanOperatingModeSchedule(always_on)
-        air_to_air_heatpump.addToNode(supply_inlet_node)
-      else
-        fan.addToNode(supply_inlet_node)
-        htg_coil.addToNode(supply_inlet_node)
-        clg_coil.addToNode(supply_inlet_node)
-      end
-      oa_system.addToNode(supply_inlet_node)
-
-      # Add a setpoint manager single zone reheat to control the
-      # supply air temperature based on the needs of this zone
-      setpoint_mgr_single_zone_reheat = OpenStudio::Model::SetpointManagerSingleZoneReheat.new(model)
-      setpoint_mgr_single_zone_reheat.setControlZone(zone)
-      setpoint_mgr_single_zone_reheat.setMinimumSupplyAirTemperature(13)
-      setpoint_mgr_single_zone_reheat.setMaximumSupplyAirTemperature(43)
-      setpoint_mgr_single_zone_reheat.addToNode(air_loop.supplyOutletNode)
-
-      # Create a diffuser and attach the zone/diffuser pair to the air loop
-      # diffuser = OpenStudio::Model::AirTerminalSingleDuctUncontrolled.new(model,always_on)
-      diffuser = OpenStudio::Model::AirTerminalSingleDuctUncontrolled.new(model, always_on)
-      air_loop.addBranchForZone(zone, diffuser.to_StraightComponent)
-      add_zone_baseboards(baseboard_type: baseboard_type, hw_loop: hw_loop, model: model, zone: zone)
-    end # zone loop
-
-    return true
-  end
-
-  # end add_sys3_single_zone_packaged_rooftop_unit_with_baseboard_heating_single_speed
-
-  def add_sys3and8_single_zone_packaged_rooftop_unit_with_baseboard_heating_multi_speed(model,
-                                                                                        zones,
-                                                                                        boiler_fueltype,
-                                                                                        heating_coil_type,
-                                                                                        baseboard_type,
-                                                                                        hw_loop)
-    # System Type 3: PSZ-AC
-    # This measure creates:
-    # -a constant volume packaged single-zone A/C unit
-    # for each zone in the building; DX cooling with
-    # heating coil: fuel-fired or electric, depending on argument heating_coil_type
-    # heating_coil_type choices are "Electric", "Gas", "DX"
-    # zone baseboards: hot water or electric, depending on argument baseboard_type
-    # baseboard_type choices are "Hot Water" or "Electric"
-    # boiler_fueltype choices match OS choices for Boiler component fuel type, i.e.
-    # "NaturalGas","Electricity","PropaneGas","FuelOil#1","FuelOil#2","Coal","Diesel","Gasoline","OtherFuel1"
-
-    always_on = model.alwaysOnDiscreteSchedule
-
-    # TODO: Heating and cooling temperature set point schedules are set somewhere else
-    # TODO: For now fetch the schedules and use them in setting up the heat pump system
-    # TODO: Later on these schedules need to be passed on to this method
-    htg_temp_sch = nil
-    clg_temp_sch = nil
-    zones.each do |izone|
-      if izone.thermostat.is_initialized
-        zone_thermostat = izone.thermostat.get
-        if zone_thermostat.to_ThermostatSetpointDualSetpoint.is_initialized
-          dual_thermostat = zone_thermostat.to_ThermostatSetpointDualSetpoint.get
-          htg_temp_sch = dual_thermostat.heatingSetpointTemperatureSchedule.get
-          clg_temp_sch = dual_thermostat.coolingSetpointTemperatureSchedule.get
-          break
-        end
-      end
-    end
-
-    zones.each do |zone|
-      air_loop = OpenStudio::Model::AirLoopHVAC.new(model)
-
-      air_loop.setName("Sys_3_PSZ_#{zone.name}")
-
-      # When an air_loop is constructed, its constructor creates a sizing:system object
-      # the default sizing:system constructor makes a system:sizing object
-      # appropriate for a multizone VAV system
-      # this systems is a constant volume system with no VAV terminals,
-      # and therfore needs different default settings
-      air_loop_sizing = air_loop.sizingSystem # TODO units
-      air_loop_sizing.setTypeofLoadtoSizeOn('Sensible')
-      air_loop_sizing.autosizeDesignOutdoorAirFlowRate
-      air_loop_sizing.setMinimumSystemAirFlowRatio(1.0)
-      air_loop_sizing.setPreheatDesignTemperature(7.0)
-      air_loop_sizing.setPreheatDesignHumidityRatio(0.008)
-      air_loop_sizing.setPrecoolDesignTemperature(13.0)
-      air_loop_sizing.setPrecoolDesignHumidityRatio(0.008)
-      air_loop_sizing.setCentralCoolingDesignSupplyAirTemperature(13.0)
-      air_loop_sizing.setCentralHeatingDesignSupplyAirTemperature(43.0)
-      air_loop_sizing.setSizingOption('NonCoincident')
-      air_loop_sizing.setAllOutdoorAirinCooling(false)
-      air_loop_sizing.setAllOutdoorAirinHeating(false)
-      air_loop_sizing.setCentralCoolingDesignSupplyAirHumidityRatio(0.0085)
-      air_loop_sizing.setCentralHeatingDesignSupplyAirHumidityRatio(0.0080)
-      air_loop_sizing.setCoolingDesignAirFlowMethod('DesignDay')
-      air_loop_sizing.setCoolingDesignAirFlowRate(0.0)
-      air_loop_sizing.setHeatingDesignAirFlowMethod('DesignDay')
-      air_loop_sizing.setHeatingDesignAirFlowRate(0.0)
-      air_loop_sizing.setSystemOutdoorAirMethod('ZoneSum')
-
-      # Zone sizing temperature
-      sizing_zone = zone.sizingZone
-      sizing_zone.setZoneCoolingDesignSupplyAirTemperature(13.0)
-      sizing_zone.setZoneHeatingDesignSupplyAirTemperature(43.0)
-      sizing_zone.setZoneCoolingSizingFactor(1.1)
-      sizing_zone.setZoneHeatingSizingFactor(1.3)
-
-      fan = OpenStudio::Model::FanConstantVolume.new(model, always_on)
-
-      staged_thermostat = OpenStudio::Model::ZoneControlThermostatStagedDualSetpoint.new(model)
-      staged_thermostat.setHeatingTemperatureSetpointSchedule(htg_temp_sch)
-      staged_thermostat.setNumberofHeatingStages(4)
-      staged_thermostat.setCoolingTemperatureSetpointBaseSchedule(clg_temp_sch)
-      staged_thermostat.setNumberofCoolingStages(4)
-      zone.setThermostat(staged_thermostat)
-
-      # Multi-stage gas heating coil
-      if heating_coil_type == 'Gas' || heating_coil_type == 'Electric'
-        htg_coil = OpenStudio::Model::CoilHeatingGasMultiStage.new(model)
-        htg_stage_1 = OpenStudio::Model::CoilHeatingGasMultiStageStageData.new(model)
-        htg_stage_2 = OpenStudio::Model::CoilHeatingGasMultiStageStageData.new(model)
-        htg_stage_3 = OpenStudio::Model::CoilHeatingGasMultiStageStageData.new(model)
-        htg_stage_4 = OpenStudio::Model::CoilHeatingGasMultiStageStageData.new(model)
-        if heating_coil_type == 'Gas'
-          supplemental_htg_coil = OpenStudio::Model::CoilHeatingGas.new(model, always_on)
-        elsif heating_coil_type == 'Electric'
-          supplemental_htg_coil = OpenStudio::Model::CoilHeatingElectric.new(model, always_on)
-          htg_stage_1.setNominalCapacity(0.1)
-          htg_stage_2.setNominalCapacity(0.2)
-          htg_stage_3.setNominalCapacity(0.3)
-          htg_stage_4.setNominalCapacity(0.4)
-        end
-
-        # Multi-Stage DX or Electric heating coil
-      elsif heating_coil_type == 'DX'
-        htg_coil = OpenStudio::Model::CoilHeatingDXMultiSpeed.new(model)
-        htg_stage_1 = OpenStudio::Model::CoilHeatingDXMultiSpeedStageData.new(model)
-        htg_stage_2 = OpenStudio::Model::CoilHeatingDXMultiSpeedStageData.new(model)
-        htg_stage_3 = OpenStudio::Model::CoilHeatingDXMultiSpeedStageData.new(model)
-        htg_stage_4 = OpenStudio::Model::CoilHeatingDXMultiSpeedStageData.new(model)
-        supplemental_htg_coil = OpenStudio::Model::CoilHeatingElectric.new(model, always_on)
-        sizing_zone.setZoneHeatingSizingFactor(1.3)
-        sizing_zone.setZoneCoolingSizingFactor(1.0)
-      else
-        raise("#{heating_coil_type} is not a valid heating coil type.)")
-      end
-
-      # Add stages to heating coil
-      htg_coil.addStage(htg_stage_1)
-      htg_coil.addStage(htg_stage_2)
-      htg_coil.addStage(htg_stage_3)
-      htg_coil.addStage(htg_stage_4)
-
-      # TODO: other fuel-fired heating coil types? (not available in OpenStudio/E+ - may need to play with efficiency to mimic other fuel types)
-
-      # Set up DX cooling coil
-      clg_coil = OpenStudio::Model::CoilCoolingDXMultiSpeed.new(model)
-      clg_coil.setFuelType('Electricity')
-      clg_stage_1 = OpenStudio::Model::CoilCoolingDXMultiSpeedStageData.new(model)
-      clg_stage_2 = OpenStudio::Model::CoilCoolingDXMultiSpeedStageData.new(model)
-      clg_stage_3 = OpenStudio::Model::CoilCoolingDXMultiSpeedStageData.new(model)
-      clg_stage_4 = OpenStudio::Model::CoilCoolingDXMultiSpeedStageData.new(model)
-      clg_coil.addStage(clg_stage_1)
-      clg_coil.addStage(clg_stage_2)
-      clg_coil.addStage(clg_stage_3)
-      clg_coil.addStage(clg_stage_4)
-
-      # oa_controller
-      oa_controller = OpenStudio::Model::ControllerOutdoorAir.new(model)
-      oa_controller.autosizeMinimumOutdoorAirFlowRate
-
-      # oa_system
-      oa_system = OpenStudio::Model::AirLoopHVACOutdoorAirSystem.new(model, oa_controller)
-
-      # Add the components to the air loop
-      # in order from closest to zone to furthest from zone
-      supply_inlet_node = air_loop.supplyInletNode
-
-      air_to_air_heatpump = OpenStudio::Model::AirLoopHVACUnitaryHeatPumpAirToAirMultiSpeed.new(model, fan, htg_coil, clg_coil, supplemental_htg_coil)
-      air_to_air_heatpump.setName("#{zone.name} ASHP")
-      air_to_air_heatpump.setControllingZoneorThermostatLocation(zone)
-      air_to_air_heatpump.setSupplyAirFanOperatingModeSchedule(always_on)
-      air_to_air_heatpump.addToNode(supply_inlet_node)
-      air_to_air_heatpump.setNumberofSpeedsforHeating(4)
-      air_to_air_heatpump.setNumberofSpeedsforCooling(4)
-
-      oa_system.addToNode(supply_inlet_node)
-
-      # Create a diffuser and attach the zone/diffuser pair to the air loop
-      # diffuser = OpenStudio::Model::AirTerminalSingleDuctUncontrolled.new(model,always_on)
-      diffuser = OpenStudio::Model::AirTerminalSingleDuctUncontrolled.new(model, always_on)
-      air_loop.addBranchForZone(zone, diffuser.to_StraightComponent)
-
-      add_zone_baseboards(baseboard_type: baseboard_type, hw_loop: hw_loop, model: model, zone: zone)
-    end # zone loop
-
-    return true
-  end
-
-  # end add_sys3_single_zone_packaged_rooftop_unit_with_baseboard_heating_multi_speed
-
-  def add_sys4_single_zone_make_up_air_unit_with_baseboard_heating(model:,
-                                                                   zones:,
-                                                                   heating_coil_type:,
-                                                                   baseboard_type:,
-                                                                   hw_loop:)
-    # System Type 4: PSZ-AC
-    # This measure creates:
-    # -a constant volume packaged single-zone A/C unit
-    # for each zone in the building; DX cooling with
-    # heating coil: fuel-fired or electric, depending on argument heating_coil_type
-    # heating_coil_type choices are "Electric", "Gas"
-    # zone baseboards: hot water or electric, depending on argument baseboard_type
-    # baseboard_type choices are "Hot Water" or "Electric"
-    # boiler_fueltype choices match OS choices for Boiler component fuel type, i.e.
-    # "NaturalGas","Electricity","PropaneGas","FuelOil#1","FuelOil#2","Coal","Diesel","Gasoline","OtherFuel1"
-    # NOTE: This is the same as system type 3 (single zone make-up air unit and single zone rooftop unit are both PSZ systems)
-    # SHOULD WE COMBINE sys3 and sys4 into one script?
-
-    always_on = model.alwaysOnDiscreteSchedule
-
-    # Create a PSZ for each zone
-    # TO DO: need to apply this system to space types:
-    # (1) automotive area: repair/parking garage, fire engine room, indoor truck bay
-    # (2) supermarket/food service: food preparation with kitchen hood/vented appliance
-    # (3) warehouse area (non-refrigerated spaces)
-
-    zones.each do |zone|
-      air_loop = OpenStudio::Model::AirLoopHVAC.new(model)
-
-      air_loop.setName("Sys_4_PSZ_#{zone.name}")
-
-      # When an air_loop is constructed, its constructor creates a sizing:system object
-      # the default sizing:system constructor makes a system:sizing object
-      # appropriate for a multizone VAV system
-      # this systems is a constant volume system with no VAV terminals,
-      # and therfore needs different default settings
-      air_loop_sizing = air_loop.sizingSystem # TODO units
-      air_loop_sizing.setTypeofLoadtoSizeOn('Sensible')
-      air_loop_sizing.autosizeDesignOutdoorAirFlowRate
-      air_loop_sizing.setMinimumSystemAirFlowRatio(1.0)
-      air_loop_sizing.setPreheatDesignTemperature(7.0)
-      air_loop_sizing.setPreheatDesignHumidityRatio(0.008)
-      air_loop_sizing.setPrecoolDesignTemperature(13.0)
-      air_loop_sizing.setPrecoolDesignHumidityRatio(0.008)
-      air_loop_sizing.setCentralCoolingDesignSupplyAirTemperature(13.0)
-      air_loop_sizing.setCentralHeatingDesignSupplyAirTemperature(43.0)
-      air_loop_sizing.setSizingOption('NonCoincident')
-      air_loop_sizing.setAllOutdoorAirinCooling(false)
-      air_loop_sizing.setAllOutdoorAirinHeating(false)
-      air_loop_sizing.setCentralCoolingDesignSupplyAirHumidityRatio(0.0085)
-      air_loop_sizing.setCentralHeatingDesignSupplyAirHumidityRatio(0.0080)
-      air_loop_sizing.setCoolingDesignAirFlowMethod('DesignDay')
-      air_loop_sizing.setCoolingDesignAirFlowRate(0.0)
-      air_loop_sizing.setHeatingDesignAirFlowMethod('DesignDay')
-      air_loop_sizing.setHeatingDesignAirFlowRate(0.0)
-      air_loop_sizing.setSystemOutdoorAirMethod('ZoneSum')
-
-      # Zone sizing temperature
-      sizing_zone = zone.sizingZone
-      sizing_zone.setZoneCoolingDesignSupplyAirTemperature(13.0)
-      sizing_zone.setZoneHeatingDesignSupplyAirTemperature(43.0)
-      sizing_zone.setZoneCoolingSizingFactor(1.1)
-      sizing_zone.setZoneHeatingSizingFactor(1.3)
-
-      fan = OpenStudio::Model::FanConstantVolume.new(model, always_on)
-
-      if heating_coil_type == 'Electric' # electric coil
-        htg_coil = OpenStudio::Model::CoilHeatingElectric.new(model, always_on)
-      end
-
-      if heating_coil_type == 'Gas'
-        htg_coil = OpenStudio::Model::CoilHeatingGas.new(model, always_on)
-      end
-
-      # TO DO: other fuel-fired heating coil types? (not available in OpenStudio/E+ - may need to play with efficiency to mimic other fuel types)
-
-      # Set up DX coil with NECB performance curve characteristics;
-
-      clg_coil = self.add_onespeed_DX_coil(model, always_on)
-
-      # oa_controller
-      oa_controller = OpenStudio::Model::ControllerOutdoorAir.new(model)
-      oa_controller.autosizeMinimumOutdoorAirFlowRate
-
-      # oa_system
-      oa_system = OpenStudio::Model::AirLoopHVACOutdoorAirSystem.new(model, oa_controller)
-
-      # Add the components to the air loop
-      # in order from closest to zone to furthest from zone
-      supply_inlet_node = air_loop.supplyInletNode
-      fan.addToNode(supply_inlet_node)
-      htg_coil.addToNode(supply_inlet_node)
-      clg_coil.addToNode(supply_inlet_node)
-      oa_system.addToNode(supply_inlet_node)
-
-      # Add a setpoint manager single zone reheat to control the
-      # supply air temperature based on the needs of this zone
-      setpoint_mgr_single_zone_reheat = OpenStudio::Model::SetpointManagerSingleZoneReheat.new(model)
-      setpoint_mgr_single_zone_reheat.setControlZone(zone)
-      setpoint_mgr_single_zone_reheat.setMinimumSupplyAirTemperature(13.0)
-      setpoint_mgr_single_zone_reheat.setMaximumSupplyAirTemperature(43.0)
-      setpoint_mgr_single_zone_reheat.addToNode(air_loop.supplyOutletNode)
-
-      # Create sensible heat exchanger
-      #              heat_exchanger = BTAP::Resources::HVAC::Plant::add_hrv(model)
-      #              heat_exchanger.setSensibleEffectivenessat100HeatingAirFlow(0.5)
-      #              heat_exchanger.setSensibleEffectivenessat75HeatingAirFlow(0.5)
-      #              heat_exchanger.setSensibleEffectivenessat100CoolingAirFlow(0.5)
-      #              heat_exchanger.setSensibleEffectivenessat75CoolingAirFlow(0.5)
-      #              heat_exchanger.setLatentEffectivenessat100HeatingAirFlow(0.0)
-      #              heat_exchanger.setLatentEffectivenessat75HeatingAirFlow(0.0)
-      #              heat_exchanger.setLatentEffectivenessat100CoolingAirFlow(0.0)
-      #              heat_exchanger.setLatentEffectivenessat75CoolingAirFlow(0.0)
-      #              heat_exchanger.setSupplyAirOutletTemperatureControl(false)
-      #
-      #              Connect heat exchanger
-      #              oa_node = oa_system.outboardOANode
-      #              heat_exchanger.addToNode(oa_node.get)
-
-      # Create a diffuser and attach the zone/diffuser pair to the air loop
-      # diffuser = OpenStudio::Model::AirTerminalSingleDuctUncontrolled.new(model,always_on)
-      diffuser = OpenStudio::Model::AirTerminalSingleDuctUncontrolled.new(model, always_on)
-      air_loop.addBranchForZone(zone, diffuser.to_StraightComponent)
-      add_zone_baseboards(baseboard_type: baseboard_type, hw_loop: hw_loop, model: model, zone: zone)
-    end # zone loop
-
-    return true
-  end
-
-  # end add_sys4_single_zone_make_up_air_unit_with_baseboard_heating
-
-  def add_sys6_multi_zone_built_up_system_with_baseboard_heating(model:,
-                                                                 zones:,
-                                                                 heating_coil_type:,
-                                                                 baseboard_type:,
-                                                                 chiller_type:,
-                                                                 fan_type:,
-                                                                 hw_loop:)
-    # System Type 6: VAV w/ Reheat
-    # This measure creates:
-    # a single hot water loop with a natural gas or electric boiler or for the building
-    # a single chilled water loop with water cooled chiller for the building
-    # a single condenser water loop for heat rejection from the chiller
-    # a VAV system w/ hot water or electric heating, chilled water cooling, and
-    # hot water or electric reheat for each story of the building
-    # Arguments:
-    # "boiler_fueltype" choices match OS choices for boiler fuel type:
-    # "NaturalGas","Electricity","PropaneGas","FuelOil#1","FuelOil#2","Coal","Diesel","Gasoline","OtherFuel1"
-    # "heating_coil_type": "Electric" or "Hot Water"
-    # "baseboard_type": "Electric" and "Hot Water"
-    # "chiller_type": "Scroll";"Centrifugal";""Screw";"Reciprocating"
-    # "fan_type": "AF_or_BI_rdg_fancurve";"AF_or_BI_inletvanes";"fc_inletvanes";"var_speed_drive"
-
-    always_on = model.alwaysOnDiscreteSchedule
-
-    # Chilled Water Plant
-
-    chw_loop = OpenStudio::Model::PlantLoop.new(model)
-    chiller1, chiller2 = self.setup_chw_loop_with_components(model, chw_loop, chiller_type)
-
-    # Condenser System
-
-    cw_loop = OpenStudio::Model::PlantLoop.new(model)
-    ctower = self.setup_cw_loop_with_components(model, cw_loop, chiller1, chiller2)
-
-    # Make a Packaged VAV w/ PFP Boxes for each story of the building
-    model.getBuildingStorys.sort.each do |story|
-      unless (BTAP::Geometry::BuildingStoreys.get_zones_from_storey(story) & zones).empty?
-
-        air_loop = OpenStudio::Model::AirLoopHVAC.new(model)
-        air_loop.setName('Sys_6_VAV with Reheat')
-        sizing_system = air_loop.sizingSystem
-        sizing_system.setCentralCoolingDesignSupplyAirTemperature(13.0)
-        sizing_system.setCentralHeatingDesignSupplyAirTemperature(13.1)
-        sizing_system.autosizeDesignOutdoorAirFlowRate
-        sizing_system.setMinimumSystemAirFlowRatio(0.3)
-        sizing_system.setPreheatDesignTemperature(7.0)
-        sizing_system.setPreheatDesignHumidityRatio(0.008)
-        sizing_system.setPrecoolDesignTemperature(13.0)
-        sizing_system.setPrecoolDesignHumidityRatio(0.008)
-        sizing_system.setSizingOption('NonCoincident')
-        sizing_system.setAllOutdoorAirinCooling(false)
-        sizing_system.setAllOutdoorAirinHeating(false)
-        sizing_system.setCentralCoolingDesignSupplyAirHumidityRatio(0.0085)
-        sizing_system.setCentralHeatingDesignSupplyAirHumidityRatio(0.0080)
-        sizing_system.setCoolingDesignAirFlowMethod('DesignDay')
-        sizing_system.setCoolingDesignAirFlowRate(0.0)
-        sizing_system.setHeatingDesignAirFlowMethod('DesignDay')
-        sizing_system.setHeatingDesignAirFlowRate(0.0)
-        sizing_system.setSystemOutdoorAirMethod('ZoneSum')
-
-        supply_fan = OpenStudio::Model::FanVariableVolume.new(model, always_on)
-        supply_fan.setName('Sys6 Supply Fan')
-        return_fan = OpenStudio::Model::FanVariableVolume.new(model, always_on)
-        return_fan.setName('Sys6 Return Fan')
-
-        if heating_coil_type == 'Hot Water'
-          htg_coil = OpenStudio::Model::CoilHeatingWater.new(model, always_on)
-          hw_loop.addDemandBranchForComponent(htg_coil)
-        end
-        if heating_coil_type == 'Electric'
-          htg_coil = OpenStudio::Model::CoilHeatingElectric.new(model, always_on)
-        end
-
-        clg_coil = OpenStudio::Model::CoilCoolingWater.new(model, always_on)
-        chw_loop.addDemandBranchForComponent(clg_coil)
-
-        oa_controller = OpenStudio::Model::ControllerOutdoorAir.new(model)
-        oa_controller.autosizeMinimumOutdoorAirFlowRate
-
-        oa_system = OpenStudio::Model::AirLoopHVACOutdoorAirSystem.new(model, oa_controller)
-
-        # Add the components to the air loop
-        # in order from closest to zone to furthest from zone
-        supply_inlet_node = air_loop.supplyInletNode
-        supply_outlet_node = air_loop.supplyOutletNode
-        supply_fan.addToNode(supply_inlet_node)
-        htg_coil.addToNode(supply_inlet_node)
-        clg_coil.addToNode(supply_inlet_node)
-        oa_system.addToNode(supply_inlet_node)
-        returnAirNode = oa_system.returnAirModelObject.get.to_Node.get
-        return_fan.addToNode(returnAirNode)
-
-        # Add a setpoint manager to control the
-        # supply air to a constant temperature
-        sat_c = 13.0
-        sat_sch = OpenStudio::Model::ScheduleRuleset.new(model)
-        sat_sch.setName('Supply Air Temp')
-        sat_sch.defaultDaySchedule.setName('Supply Air Temp Default')
-        sat_sch.defaultDaySchedule.addValue(OpenStudio::Time.new(0, 24, 0, 0), sat_c)
-        sat_stpt_manager = OpenStudio::Model::SetpointManagerScheduled.new(model, sat_sch)
-        sat_stpt_manager.addToNode(supply_outlet_node)
-
-        # Make a VAV terminal with HW reheat for each zone on this story that is in intersection with the zones array.
-        # and hook the reheat coil to the HW loop
-        (BTAP::Geometry::BuildingStoreys.get_zones_from_storey(story) & zones).each do |zone|
-          # Zone sizing parameters
-          sizing_zone = zone.sizingZone
-          sizing_zone.setZoneCoolingDesignSupplyAirTemperature(13.0)
-          sizing_zone.setZoneHeatingDesignSupplyAirTemperature(43.0)
-          sizing_zone.setZoneCoolingSizingFactor(1.1)
-          sizing_zone.setZoneHeatingSizingFactor(1.3)
-
-          if heating_coil_type == 'Hot Water'
-            reheat_coil = OpenStudio::Model::CoilHeatingWater.new(model, always_on)
-            hw_loop.addDemandBranchForComponent(reheat_coil)
-          elsif heating_coil_type == 'Electric'
-            reheat_coil = OpenStudio::Model::CoilHeatingElectric.new(model, always_on)
-          end
-
-          vav_terminal = OpenStudio::Model::AirTerminalSingleDuctVAVReheat.new(model, always_on, reheat_coil)
-          air_loop.addBranchForZone(zone, vav_terminal.to_StraightComponent)
-          # NECB2011 minimum zone airflow setting
-          min_flow_rate = 0.002 * zone.floorArea
-          vav_terminal.setFixedMinimumAirFlowRate(min_flow_rate)
-          vav_terminal.setMaximumReheatAirTemperature(43.0)
-          vav_terminal.setDamperHeatingAction('Normal')
-
-          # Set zone baseboards
-          add_zone_baseboards(model: model,
-                              zone: zone,
-                              baseboard_type: baseboard_type,
-                              hw_loop: hw_loop)
-        end
-      end
-    end # next story
-
-    # for debugging
-    # puts "end add_sys6_multi_zone_built_up_with_baseboard_heating"
-
-    return true
-  end
 
   def setup_hw_loop_with_components(model,
                                     hw_loop,
@@ -3325,6 +2234,73 @@ class NECB2011
                                                                          clg_coil)
     ptac.setName("#{zone.name} PTAC")
     ptac.addToThermalZone(zone)
+  end
+
+  def common_air_loop(model: model)
+    mau_air_loop = OpenStudio::Model::AirLoopHVAC.new(model)
+    air_loop_sizing = mau_air_loop.sizingSystem
+    air_loop_sizing.autosizeDesignOutdoorAirFlowRate
+    air_loop_sizing.setMinimumSystemAirFlowRatio(1.0)
+    air_loop_sizing.setPreheatDesignTemperature(7.0)
+    air_loop_sizing.setPreheatDesignHumidityRatio(0.008)
+    air_loop_sizing.setPrecoolDesignTemperature(13.0)
+    air_loop_sizing.setPrecoolDesignHumidityRatio(0.008)
+    air_loop_sizing.setSizingOption('NonCoincident')
+    air_loop_sizing.setCoolingDesignAirFlowMethod('DesignDay')
+    air_loop_sizing.setCoolingDesignAirFlowRate(0.0)
+    air_loop_sizing.setHeatingDesignAirFlowMethod('DesignDay')
+    air_loop_sizing.setHeatingDesignAirFlowRate(0.0)
+    air_loop_sizing.setSystemOutdoorAirMethod('ZoneSum')
+    air_loop_sizing.setCentralCoolingDesignSupplyAirHumidityRatio(0.0085)
+    air_loop_sizing.setCentralHeatingDesignSupplyAirHumidityRatio(0.0080)
+    return mau_air_loop
+  end
+
+  def create_heating_cooling_on_off_availability_schedule(model)
+    #todo Create a feature to derive start and end heating and cooling seasons from weather file.
+    avail_data = [{start_month: 1, start_day: 1, end_month: 6, end_day: 30, htg_value: 1, clg_value: 0},
+                  {start_month: 7, start_day: 1, end_month: 10, end_day: 31, htg_value: 0, clg_value: 1},
+                  {start_month: 11, start_day: 1, end_month: 12, end_day: 31, htg_value: 1, clg_value: 0}
+    ]
+
+    # Heating coil availability schedule for tpfc
+    htg_availability_sch = OpenStudio::Model::ScheduleRuleset.new(model)
+    htg_availability_sch.setName('tpfc_htg_availability')
+    # Cooling coil availability schedule for tpfc
+    clg_availability_sch = OpenStudio::Model::ScheduleRuleset.new(model)
+    clg_availability_sch.setName('tpfc_clg_availability')
+    avail_data.each do |data|
+      htg_availability_sch_rule = OpenStudio::Model::ScheduleRule.new(htg_availability_sch)
+      htg_availability_sch_rule.setName('tpfc_htg_availability_sch_rule')
+      htg_availability_sch_rule.setStartDate(model.getYearDescription.makeDate(data[:start_month], data[:start_day]))
+      htg_availability_sch_rule.setEndDate(model.getYearDescription.makeDate(data[:end_month], data[:end_day]))
+      htg_availability_sch_rule.setApplySunday(true)
+      htg_availability_sch_rule.setApplyMonday(true)
+      htg_availability_sch_rule.setApplyTuesday(true)
+      htg_availability_sch_rule.setApplyWednesday(true)
+      htg_availability_sch_rule.setApplyThursday(true)
+      htg_availability_sch_rule.setApplyFriday(true)
+      htg_availability_sch_rule.setApplySaturday(true)
+      day_schedule = htg_availability_sch_rule.daySchedule
+      day_schedule.setName('tpfc_htg_availability_sch_rule_day')
+      day_schedule.addValue(OpenStudio::Time.new(0, 24, 0, 0), data[:htg_value])
+
+      clg_availability_sch_rule = OpenStudio::Model::ScheduleRule.new(clg_availability_sch)
+      clg_availability_sch_rule.setName('tpfc_clg_availability_sch_rule')
+      clg_availability_sch_rule.setStartDate(model.getYearDescription.makeDate(data[:start_month], data[:start_day]))
+      clg_availability_sch_rule.setEndDate(model.getYearDescription.makeDate(data[:end_month], data[:end_day]))
+      clg_availability_sch_rule.setApplySunday(true)
+      clg_availability_sch_rule.setApplyMonday(true)
+      clg_availability_sch_rule.setApplyTuesday(true)
+      clg_availability_sch_rule.setApplyWednesday(true)
+      clg_availability_sch_rule.setApplyThursday(true)
+      clg_availability_sch_rule.setApplyFriday(true)
+      clg_availability_sch_rule.setApplySaturday(true)
+      day_schedule = clg_availability_sch_rule.daySchedule
+      day_schedule.setName('tpfc_clg_availability_sch_rule_day')
+      day_schedule.addValue(OpenStudio::Time.new(0, 24, 0, 0), data[:clg_value])
+    end
+    return clg_availability_sch, htg_availability_sch
   end
